QA-Python-homework-02 review fixes
